Feature sergey lyubimov

+++ src/main/java/ru/skypro/homework/dtos/AdsDto.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 @Data
 public class AdsDto {
 
-    private Long author;
+    private Integer author;
     private String image;
     private Long pk;
     private BigDecimal price;
